Use pristine types as a bottleneck for dropping memory references
Sometimes crawler runs can go for hours and visit many pages. Each page can be large but we only need a small amount of information from it, and it's important not to hold memory references to the original html to avoid bloat. The current solution for this is manually inserted deep copy calls, which work for now but won't be maintainable with future crawler changes.

This change mandates that the webpage output interface uses special "pristine" types for strings and URLs, which copy the memory when constructed. This way the type system will enforce the maintainable choices rather than us having to remember this quirk.

Alternative designs considered:
* Copy every piece of data that's extracted from html right away - heuristic matching has a lot of churn with small strings and copying consumes even more memory, upfront
* Split the string and URL types into "clean" and "dirty" throughout the crawler code - this would touch a lot of code and require a lot of acrobatics with generics, and we're really only concerned that the webpage output is clean.